&&演算子の抜けを修正

+++ docker-entrypoint.sh
@@ -281,7 +281,7 @@ EOPHP
 	done
 fi
 
-if [ -e /var/www/html/wordpress/index.php ] && [ -e /var/www/html/wordpress/.htaccess ] [ ! -e /var/www/html/index.php ] && [ ! -e /var/www/html/.htaccess ]; then
+if [ -e /var/www/html/wordpress/index.php ] && [ -e /var/www/html/wordpress/.htaccess ] && [ ! -e /var/www/html/index.php ] && [ ! -e /var/www/html/.htaccess ]; then
   sed -i -e 's/\/wp-blog-header\.php/\/wordpress\/wp-blog-header.php/g' /var/www/html/wordpress/index.php
   mv /var/www/html/wordpress/.htaccess /var/www/html/.htaccess
   mv /var/www/html/wordpress/index.php /var/www/html/index.php
